# Parsing optimisations