Make low-level parsing more closely aligned to what we need, to save memory allocations.
See individual commit messages for more details.

These routines are called a lot, so this change reduces garbage collection

Benchmarks before (c5bdebdffebfa7cc0839770a3bf70429260f1bc9):
```
BenchmarkTopologyList-2         	       2	 603197223 ns/op	112161732 B/op	 1253503 allocs/op
BenchmarkTopologyHosts-2        	       3	 385415232 ns/op	82957173 B/op	 1015204 allocs/op
BenchmarkTopologyContainers-2   	      10	 192766565 ns/op	33145659 B/op	  407698 allocs/op
```

after this PR:
```
BenchmarkTopologyList-2         	       3	 484253544 ns/op	104023597 B/op	  977911 allocs/op
BenchmarkTopologyHosts-2        	       3	 366672131 ns/op	75139373 B/op	  742471 allocs/op
BenchmarkTopologyContainers-2   	      10	 178938068 ns/op	29289833 B/op	  272342 allocs/op
```